Extracting FileReader class and renaming
