+++ C# Star Practices/StudentsDataManager/Models/Lessons.cs
@@ -1,4 +1,4 @@
-namespace StudentsDataManager.Objects;
+namespace StudentsDataManager.Models;
 
 public class Lesson
 {
